fix: 500 при передаче акта клиенту — явный каст $4::int в CASE
Postgres не мог вывести тип параметра $4 внутри CASE WHEN ... THEN $4
ELSE NULL END и слал его как text, хотя колонка shared_with_client_by
integer. Явный каст решает 42804 column is of type integer but
expression is of type text.

+++ server/src/modules/acts/acts.service.js
@@ -172,7 +172,7 @@ async function setActShared({ tenantId, actId, userId, shared }) {
   const r = await query(
     `UPDATE wms.acceptance_acts
      SET shared_with_client_at = CASE WHEN $3 THEN NOW() ELSE NULL END,
-         shared_with_client_by = CASE WHEN $3 THEN $4 ELSE NULL END,
+         shared_with_client_by = CASE WHEN $3 THEN $4::int ELSE NULL END,
          updated_at = NOW()
      WHERE id=$1 AND tenant_id=$2
      RETURNING *`,
